requirements.txt
absl-py==0.11.0
astunparse==1.6.3
attrs==20.3.0
backcall==0.2.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.6.0
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
dm-tree==0.1.5
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
gpflow==2.1.4
grpcio==1.32.0
h5py==2.10.0
idna==2.10
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.35.0
Markdown==3.3.3
matplotlib==3.3.3
multipledispatch==0.6.0
numba==0.52.0
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.8
pandas==1.2.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.14
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.4
pykalman==0.9.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.6.0
six==1.15.0
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-probability==0.12.1
termcolor==1.1.0
tornado==6.1
tqdm==4.56.2
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
